requirements.txt
huggingface_hub
google-generativeai
python-dotenv